[x] make the test-only config a list of package names
Also check that the package names actually exist. Whoops, this was a
mistake on my part to not check if package names exist.

There is one semantic change: we were missing move-vm-test-utils, and
diem-transaction-replay depends on it, so I moved move-vm-test-utils to
the test-only section. This caused diem-validator-interface to also be
marked as test-only.