[move prover] build a GlobalEnv from bytecodes
In order to deploy the read/write set analysis, the Diem adapter will need to build a GlobalEnv from raw bytecodes stored on-chain. This PR makes that possible by adding a new builder for the GlobalEnv that needs only a list of modules topologically sorted by the module dependency relation.

To test the change, I run the bytecode builder on the modules extracted from each GlobalEnv in the existing test harness + do some sanity checks to ensure that both environments contain the same code.

Closes: #7318