[move-prover][refactoring] Use indices instead of symbols to refer to temporaries, implement specs in code.
This PR provides a long-wanted refactoring to eliminate the symbolic reference to bytecode temporaries by index-based references. The new exp variant `Temporary(id, index)` serve for this purpose. This removes a long-standing bug regards name shadowing of temporaries which the prover cannot correctly resolve based on bytecode only. Also, spec blocks in code are now using the indices provided by the Move compiler instead of symbolic lookup. This allows to remove a number of TODO's and simplifications.

This PR also enables spec blocks in code for the V2 compilation scheme which are now reduced to `Bytecode::Prop(..)` instructions at time of stackless bytecode generation.

After this PR, we are ready to re-visit loop invariants for the V2 compilation scheme.

A lot of baseline diffs in the bytecode crate stem from that temporaries are not longer symbolically displayed but by index.

Closes: #7372