[move-prover][refactoring] Implement modifies for V2 and port tests.
This implements the `modifies` clause in specifications for the new compilation scheme. The related tests are ported over to V2.

Modifies is handled by introducing a new expression `CanModify<R<T>>(addr)` which is asserted at the appropriate points in the bytecode. In contrast to the old implementation, those assertions are only generated for the top-level verification entry, assuming called functions are already verified. For opaque calls, the instruction `modifies global<R<T>>(addr)` is translated to havoc the memory at the dedicated `addr`, using the same trick `M' == M[addr := M'[addr]]` as in the old implementation.

In the new translation scheme, it appears we do not to pass the caller-domain parameter any longer, as we are checking modifies at caller side.

Closes: #7400