client: introduce new unified client
Currently there are 4 different implementations of diem json rpc clients
(2 sync and 2 async) using 3 different sets of types for the wire
format. Each of these different clients made various design decisions,
usually intended to solve a single internal problem, and none of them
were designed with the intention to have a stable API to be used
external from diem-core.

In an effort to begin putting together a stable rust-sdk intended for
external use, introduce a new diem json rpc client library. This new
implementation is designed with the intent to have a stable API and
will hopefully be published on crates.io at some point in the near
future. It is also intended to replace the various other client
implementations found in the diem codebase in order to avoid
[xkcd #927](https://xkcd.com/927/).

This new library provides an async client as well as a fully sync client
which can be selected via the features 'async' and 'blocking'
respectively. Both of this implementations are built on a single set of
types for the wire format.