[move prover] CompositionalAnalysis::summarize returns a Summary
Previously, this function took ownership of FunctionData, mutated its `annotations` to add the summary, then handed it back. This is convenient because the caller doesn't have to remember to update the annotations, but it creates ownership problems in the case that the caller analysis wants to store a `FunctionTarget` or `FunctionEnv` as part of its analysis state (which is useful if, e.g., you want to know if a variable is a formal parameter of the current function or a local). This was creating problems for some upcoming changes to the read/write set analysis that require a `FunctionEnv` in the analysis state.

Closes: #8243