[genesis] simplify validator and operator initialization
Currently, the caller of `vm_genesis_encode_genesis_transaction` must pass two vectors of type `(Option<Ed25519PublicKey>, Name, ScriptFunction)` and `(Ed25519PublicKey, Name, ScriptFunction)`. The structure and relationship of these vectors is fairly complex. This PR simplifies the API so the caller only needs to pass a single vector of `Validator` objects that contain the necessary data about a validator and its operator.

This simplification lets us delete a lot of code in both the config builder and vm_genesis.

Closes: #8517