[move-vm] allow multi-module publication
This commit allows the Move VM to take a vector of modules (for lacking
of a better name, this vector of modules is called a bundle for now) and
publish them together.

Modules in the bundle vector must follow a topological order, i.e., if
there is a dependency relation (usage or friend) between two modules
A and B, via either 1) B uses A or 2) A declares B as a friend. Then,
A must appear before B in the bundle vector in order to be published
together. Putting B in front of A in this case will lead to a
MISSING_DEPENDENCY error or other linker errors.

All modules in the bundle will have to pass the bytecode verifier and
will be linker-checked with existing modules and all other modules in
the same bundle. Either all modules in the bundle are published/updated
or none is published/updated. Publication of the whole bundle fails as
long as one module in the bundle fail the bytecode verifier or the
linker checks (i.e., dependency and cyclic dependency check).