[move] Remove `into_module` from disassembler
Remove usages of deprecated method `CompiledScript::into_module` from
the disassembler and bytecode-source-map packages. This is accomplished
by converting members of type `CompiledModule` to `BinaryIndexedView`.
One additional change here is that `BinaryIndexedView` only keeps a
reference to its underlying module or script, so the disassembler and
source maps no longer own the module or script data they represent, and
so lifetime markers are introduced to ensure that the underlying
resource is not freed prematurely.

Closes: #8629