[move-prover] Invariant refactoring: Eliminate TypeLocal.
This PR eliminates the redundant type variant of `TypeLocal`. This type variant was needed for type quantification, but as we embrace full monomorphization, we remove this feature, and instead replace it with generic invariants. (As in `invariant<T> P`.) Before this PR, we had used TypeLocal as a convenience to distinguish type parameters of invariants from type paramneters of functions, but this created boilerplate and made the code mor complicated. In this PR, we exclusively use the existing concept of `TypeParameter`.

This PR aims to maintain the current design of global invariant injection as much as possible; changes to this are subject of future PRs.

In detail:

- Both local_mono and local_mono_compat processors have been removed. The computation of the instantiations of invariants is now happening at the time the invariant is injected/instrumented. This is necessary because once an invariant is injected into bytecode, we can't distinguish it's type parameters from that of the enclosing function. Therefore we need to fully instantiate invariants at injection time.
- Instantiation of invariant expressions is now handled by `SpecTranslator`. This module was already able to instantiate expressions during translation; the entrypoints for invariant translation have been modified accordingly.
- A lot of code has been removed besides `TypeLocal`. The quantifier elimination in `mono_analysis` is not longer needed. The axiom instantiation in the same place was likely erroneous and needs to be redone. The v1 of `global_invariant_instrumentation and `usage_analysis` has been removed to avoid refactoring them.
- The diem framework and move stdlib has been updated to use the new generic invariant semantics.
- As a side effect, also adding a new script to validate a PR locally in `./tools/check_pr.sh`.

Closes: #8946