[move-cli] deep copy tests into the temp dir before running it there
The prior fix #8938 only creates a temp dir for the test but the actual
Move commands are still executed in the test case folder (which is
scanned by the datatest). Hence, we still see the flaky issue #8917.

This commit copies everything in the test case folder to the temp dir
first and then execute the test case there.

Closes: #8984