[move] Remove `OwnedLoc` in favor of just `Loc`
The Move bytecode-source-map module defines an `OwnedLoc` type that is
meant to retain a heap-allocated `String` filename. This was once
necessary because the "source location" type defined in `move-ir/types`,
`Loc`, stored its `file` field as a static `&str` reference, and so did
not "own" its data.

The implementation of `Loc` has since changed to use an interned string,
`move_symbol_pool::Symbol`. `Symbol` objects are stored globally and can
be serialized and deserialized just as `String` objects can be. As a
result, there is no longer a good reason to ever use `OwnedLoc` instead
of `Loc`.

This commit deletes `OwnedLoc` without touching any of the code
dedicated to mapping between `OwnedLoc` and `Loc`.