[move-prover][paper] Extracting paper from tech report.
This PR extracts a paper version out of the tech report. It moves some content into appendices, and simplifies and unifies other. This way, we can describe the core aspects of memory model, global invariants, and monomorphization within 16 pages + appendices. Not much of the content of the original report has been dropped (rather gone in dependencies), so we can consider dropping the report and continue work with this.

There are some changes here (specifically to new text about global invariants) which require further discussion, which better happens in person.

We have some 3-4 pages left for discussion and other content we want to add.

Closes: #9238