[debug-endpoint]: add api for dumping config + git revision
Adding a /node-info route to collect data about the node_config and git_revision from a given node. NodeDebugService now takes in a reference to a NodeConfig as a parameter. Created a NodeInfo struct to contain the git_revision and node_config fields and make it easy to serialize the data using Serde macros. Also added dependencies to cargo.toml to support serde serialization and updated references to NodeDebugService in other files like in diem-node > src > lib.rs.

Closes: #9376