[move-prover][paper] elaborated the memory model a bit
The majority of the editing comes from more elaboration of
the memory model (in particular, the borrow semantics).

Additions include
- Vector::borrow, Vector::borrow_mut to create referencs
- How to enforce the borrow semantics for global memories (via `acquires`)
- The life time of borrows from global locations

Unfortunately, this pushes the paper to 17 pages. So I did some
rephrasing on paragraphs that ends with a line that has only
a single word... Now the paper is back to 16 pages again.

This commit also includes other editing suggestions from David,
including
- Added two references on hacks on smart contracts
- Updated ack with a former intern

Closes: #9402