[move-prover][paper] Refinements
- Adding the changes Emma suggested on Slack
- Writing a better introduction to Monomorphization and why we are doing it.
- Cut the invariant failure examples in the move section to make space for the above.

I also tried to fix the problem with the overfull boxes for the URLs added for Medium articles for bugs in blockchains, but it appears we eather don't have hyperlinks via hyperref package and use url package with hyphens option, or live with those boxes. I think for now the later is better.

Closes: #9406