[decoupled-execution] adjust the sync condition
This commit simplifies the sync condition to avoid race condition between sync and buffer manager.

Before it's possible for a node with version X to trigger a sync to X+10 while buffer manager locally commits to X+15.
After it's not possible to commit beyond the sync target since it's only triggered if the block is not in buffer manager yet.

Closes: #9529