Enabled the tests of FixedPoint32
Previously, the behavior of FixedPoint32 was abstracted away, and the
tests of FixedPoint32 were disabled because Z3 had an issue of solving
non-linear integer arithmetic (NIA).

In the recent releases, the NIA solver in Z3 was reimplemented, and the
performance has improved, so Prover can reason about something related
to FixedPoint32.

This PR does the followings:
- add some test cases for non-linear integer arithmetic
- unabstract the FixedPoint32 behavior
- enable the tests for the FixedPoint32

The performance impact of "unabstracting" the FixedPoint32 behavior
seems to be negligible. Here are the running times of `cargo test -p
move-prover --release` before and after the change.

- before the change (abstracting the FixedPoint32 behavior)
  - (tested 3 times) 223 secs, 230 secs, 253 secs
- after the change (without abstracting the FixedPoint32 behavior)
  - (tested 3 times) 235 secs, 243 secs, 250 secs

The Diem Framework does not heavily use the FixedPoin32 module. It is
used only in the following places:
- The modules such as `XUS`, `XDX` and `TreasuryComplianceScripts`
  create and assign a FixedPoint32 number.
- The `Diem` module uses the FixedPoint32 multipication one time.

Closes: #9719