[consensus] pass pending_txns properly with decoupled execution
pending_txns is used to decide if we want to wait for available txns from mempool or proceed with empty blocks.

it uses the exclude_payload to determine which is all blocks from the commit root.
after decoupled execution, it should only use all blocks from ordered root to determine if we should proceed with empty blocks.

Closes: #9580